requirements.txt
absl-py==1.4.0
anyio==3.6.2
appdirs==1.4.4
astunparse==1.6.3
bcrypt==4.0.1
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
fastapi==0.95.1
flatbuffers==23.5.26
frozendict==2.3.8
gast==0.4.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.56.0
h11==0.14.0
h5py==3.9.0
html5lib==1.1
idna==3.4
jax==0.4.13
joblib==1.3.1
keras==2.12.0
libclang==16.0.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.3
ml-dtypes==0.2.0
multitasking==0.0.11
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
protobuf==4.23.3
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
sklearn==0.0.post5
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.13
starlette==0.26.1
termcolor==2.3.0
threadpoolctl==3.1.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.22.0
webencodings==0.5.1
Werkzeug==2.3.6
wrapt==1.14.1
yfinance==0.2.18
